chore: fix conversion warnings

+++ src/am4utils/cpp/demand.cpp
@@ -5,12 +5,19 @@ PaxDemand::PaxDemand() : y(0), j(0), f(0) {}
 PaxDemand::PaxDemand(uint16_t y, uint16_t j, uint16_t f) : y(y), j(j), f(f) {}
 
 PaxDemand PaxDemand::operator/(double load) const {
-    return PaxDemand(floor(y / load), floor(j / load), floor(f / load));
+    return PaxDemand(
+        static_cast<uint16_t>(floor(static_cast<double>(y) / load)),
+        static_cast<uint16_t>(floor(static_cast<double>(j) / load)),
+        static_cast<uint16_t>(floor(static_cast<double>(f) / load))
+    );
 }
 
 CargoDemand::CargoDemand() : l(0), h(0) {}
 CargoDemand::CargoDemand(uint32_t l, uint32_t h) : l(l), h(h) {}
-CargoDemand::CargoDemand(const PaxDemand& pax_demand) : l(round(pax_demand.y / 2.0F) * 1000), h(pax_demand.j * 1000) {}
+CargoDemand::CargoDemand(const PaxDemand& pax_demand) :
+    l(static_cast<uint32_t>(round(pax_demand.y / 2.0) * 1000)),
+    h(pax_demand.j * 1000)
+{}
 
 
 const string PaxDemand::repr(const PaxDemand& demand) {
